(1): bug fixes for Statistic and Trials
(2): User now notified by Email when Experiment is successful.

(3): Layout update for statistics.

(4): Minimum Trial Logic now fixed.

(5): User can now no longer participate in their own experiments & also when its ended. However, can still view Statistics for given trials.

+++ app/src/main/java/com/example/nerdherd/NonnegativeTrial.java
@@ -16,17 +16,20 @@ import java.util.ArrayList;
 public class NonnegativeTrial extends Trial {
 
     private Integer totalCount;
-    private ArrayList<Integer> nonNegativeTrials;
+    private ArrayList<Long> nonNegativeTrials;
 
-    public NonnegativeTrial(ArrayList<Integer>nonNegatives){
+    public NonnegativeTrial(ArrayList<Long>nonNegatives){
         this.nonNegativeTrials = nonNegatives;
     }
 
-    public ArrayList<Integer> getNonNegativeTrials(){
+    public ArrayList<Long> getNonNegativeTrials(){
+
+//
         return nonNegativeTrials;
     }
 
-    public void setNonNegativeTrials(ArrayList<Integer> nonNegativeResult){
+
+    public void setNonNegativeTrials(ArrayList<Long> nonNegativeResult){
         this.nonNegativeTrials = nonNegativeResult;
     }
 }
